convert grpc code to int

+++ errors.go
@@ -59,7 +59,7 @@ func SetSpanTags(span opentracing.Span, err error, client bool) {
 		code = s.Code()
 	}
 	span.SetTag("response_code", code)
-	span.SetTag("status.code", code)
+	span.SetTag("status.code", uint32(code))
 	span.SetTag("response_class", c)
 	if err == nil {
 		return
